backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
idna==3.10
numpy==2.2.1
oauthlib==3.2.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
